rename example application to client_echo
